requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
redis
python-slugify
qrcode[pil]
python-dotenv
fakeredis
ruff
pytest
pytest-asyncio
httpx
python-multipart
Pillow
